requirements.txt
numpy>=1.26
opencv-python>=4.9
pillow>=10.2
pyyaml>=6.0
tqdm>=4.66

# fast-path: open-vocabulary detector
ultralytics>=8.3.0          # YOLOE / YOLO-World / YOLOv10 family

# slow-path: VLM verifier (Cosmos-Reason2-2B is Qwen3-VL architecture)
transformers>=4.46
accelerate>=0.34
safetensors>=0.4
qwen-vl-utils>=0.0.10       # video frame packing for Qwen3-VL chat templates
av>=12.0                    # PyAV for video reading inside qwen-vl-utils

# optional, only required for configs/tiny.yaml (Cosmos in W4A16 / NF4):
# bitsandbytes>=0.43
